Fix vendor buy exploit, chat commands, defrag search and pathfinding
- Buying accepted any vendable item in the world, not just the vendor's
  stock, letting a crafted packet copy or take other players' items. The
  cost total could also wrap before the limit check.
- Chat commands kept only the last word of their argument, /whereis with
  no name crashed, and it reported the wrong player.
- The -G defrag UID search missed most UIDs (corrupting links) and could
  loop forever; use a proper binary search and bound the UID table.
- A* relaxed nodes with the wrong cost, and re-sorted the open list on
  every step; pick the cheapest node with one scan instead.
- addPlayerSee ran CanSee and several region lookups per nearby item on
  every step; do the cheap distance tests first and look up once.

+++ src/graysvr/CClientEvent.cpp
@@ -1081,7 +1081,7 @@ void CClient::Event_VendorBuy(CChar* pVendor, const VendorItem* items, size_t it
 			return;
 		}
 
-		costtotal += (items[i].m_amount * items[i].m_price);
+		costtotal += static_cast<INT64>(items[i].m_amount) * items[i].m_price;
 		if ( costtotal > MAX_COST )
 		{
 			pVendor->Speak("Alas, I am not allowed to operate such huge sums.");
